+++ src/calibrated_explanations/calibration/state.py
@@ -13,7 +13,7 @@ from typing import TYPE_CHECKING, Any
 
 import numpy as np
 
-from calibrated_explanations.core import DataShapeError
+from ..utils.exceptions import DataShapeError
 
 from ..utils import safe_isinstance